In all of the response handlers that idle the connection if there's nothing left in the queue, don't idle if there are writes pending.  The cases where this could possibly trigger are obscure and involve the remote peer doing evil things, but the rest of the code handles it correctly and we were still seeing assertion failures, indicating that evil may be happening.
In issueStreamingCommands, make certain that there are no pending writes
before idling the connection.

Add the code to ihaveBodyDone that was already in commandWriteDone to idle
the connection if the queue is empty in case we'd had to defer the idle in
the response handler due to an unfinished write.

Whenever doSomeWrites is called with writes still pending, add a work
callback to do the write at the next opportunity.  This should eliminate a
temporary connection deadlock state on flushing, where the response to the
IHAVE body arrived before we finished writing it.  Before, doSomeWrites
would have failed to call issueQUIT because writes were still pending, and
then after the writes complete, there's no code to go back and issue it
until the read timeout expires.


git-svn-id: http://inn.eyrie.org/svn/trunk@7382 38acf0ee-cff8-0310-b87c-e736704562cb