Add an overview_cancel (which takes group and article number) and overview_cancel_xref (which takes a token and retrieves the Xref header from the article to find groups and article numbers) interfaces to the new overview API.
Change the underlying cancel method of overview backends to take a group
and article number rather than a token, since an overview backend is
never going to know what to do with a token.

Modify OVcancel, copying the new overview_cancel_xref implementation, to
retrieve the article and parse the Xref header to find groups and article
numbers and then call the cancel method for each pair.

Implement the new cancel method for tradindexed.  buffindexed and ovdb
still have stub implementations at this point.

Modify innd to cancel the overview data for an article whenever cancelling
an article.


git-svn-id: http://inn.eyrie.org/svn/trunk@7387 38acf0ee-cff8-0310-b87c-e736704562cb