Various fixes from Heiko Schlichting for DBZTEST mode:
Remove the right file in RemoveDBZ and use concat instead of a static
buffer for filenames.  Add a command-line option to set nfswriter if
desired and avoid a core dump from a missing innconf.  Fix debugging
output from dbzclose and return checking from dbzstore.  Fix an off-by-one
error in the line counter.  Fix a typo in usage.


git-svn-id: http://inn.eyrie.org/svn/branches/2.4@7497 38acf0ee-cff8-0310-b87c-e736704562cb