docheckgroups now handles wireformat articles (for using it manually from the spool) and warns when there are missing or obsolete descriptions.
Newgroup control messages for existing groups now change their description.
If a mail is sent to administrators, it reminds them to update their
newsgroups file.
Now pretty print the newsgroups file (from one to three tabulations between
the name of the group and its short description).
Refuse the creation of to. groups.

Checkgroups when there is nothing to change no longer result in sending
a blank mail to administrators.


git-svn-id: http://inn.eyrie.org/svn/trunk@7632 38acf0ee-cff8-0310-b87c-e736704562cb