Support for "runasuser" and "runasgroup" options within inn.conf in order to set the news user and the news group under which the news server runs.
Remove all the occurrences of previous NEWSUSER and NEWSGRP
in the source code.  Instead of them, a new library "newsuser.h"
is called whenever we have to ensure that the program is running
as the right news user/group (ensure_news_* functions).

Many thanks to Ivan Shmakov for this patch.



git-svn-id: http://inn.eyrie.org/svn/trunk@7703 38acf0ee-cff8-0310-b87c-e736704562cb