In RFC 3977:
   o  If it advertises the READER capability, it MUST return a 200 or
      201 response with the same meaning as for the initial greeting; in
      this case, the command MUST NOT affect the server state in any
      way.

Therefore, we do not answer 502 to such a command.  Sorry for my
revision 7975 which did that for nnrpd after a successful authentication.
Note that the same behaviour for innd is the right thing to do
after authentication because innd advertises neither MODE-READER
nor READER after that (so my revision 7974 is right for innd).


git-svn-id: http://inn.eyrie.org/svn/trunk@7991 38acf0ee-cff8-0310-b87c-e736704562cb