news.daily:  don't send superfluous mails when "nomail" is given
Make sure that when news.daily is called with "nomail", mail is only
sent when there is real output.  Previously, there would always be
headings and empty lines useful to structure the full report, which are
now ommitted.

Also, postexec has been moved up to be included in the regular mail,
and the "nomail" definition of ${MAIL} is no longer overwritten with
the default.

Thanks to Florian Schlichting for this patch.


git-svn-id: http://inn.eyrie.org/svn/branches/2.5@8807 38acf0ee-cff8-0310-b87c-e736704562cb