When using tradindexed, the overview data for a cancelled article is now properly immediately removed from the overview.
OVcancel was failing because the article number could be higher than the
cached high water mark.  The overview data is now reopened so as to check
the actual high water mark.
The article number was not always parsed without error (a trailing "\r\n"
may be present).

Thanks to Lars Magne Ingebrigtsen for the patch.


git-svn-id: http://inn.eyrie.org/svn/branches/2.5@8966 38acf0ee-cff8-0310-b87c-e736704562cb