When HDR/XHDR/XPAT was used on a new article coming into a newsgroup, requesting a header not present in the overview database, the first subsequent OVER/XOVER command did not show that article.  A remap of the overview data file was missing in nnrpd.
Thanks to Sam Varshavchik for the bug report.

The cached index and the cached data files were not immediately remapped
to reflect the change.  When doing an overview search with (X)OVER, both
the index and data files are remapped, if need be.  The same for (X)HDR
or XPAT *when* using the overview.
Here, as "(from articles)" is replied by the server, the search is done
in the article itself and does not call an OVopensearch().  The code uses
OVgetartinfo() which calls tdx_article_entry() which remaps the index file
but *not* the data file.  Therefore, the high water mark in the cached
data file is still mapped to high-1 after (X)HDR but tdx_article_entry()
updates data->high to be the new high water mark...  Then (X)OVER is used.
It believes the cached data information is right.  But obviously does
not find the last article and the cached data file then is marked as
to-be-remapped.  Which is done when doing the subsequent overview search
with (X)OVER.


See also [8451].


git-svn-id: http://inn.eyrie.org/svn/branches/2.5@9054 38acf0ee-cff8-0310-b87c-e736704562cb