This patch causes both CRwithoutLF (\r) and LFwithoutCR (\n) to be considered the same as CRLF (\r\n) for purposes of checking allowable header line length.
It appears that some people like to grace their articles with Face: or
X-Face: headers of typically 4-6k binary data.  While usually put into
nice continuation lines of around 80 characters, a few clients fail to
create proper line breaks for this header, inserting only a single \r or
\n instead.  The nnrpd in INN 2.3 doesn't reject or correct these mistakes
for non-system headers, and offending articles are still "in the wild".

innd currently counts, logs and ignores such imperfect line breaks.
Alas, when checking for the maximum allowable 'physical' line length,
single \r or \n are not recognized, thus making the whole 'virtual'
(continued) header appear as a single line, easily exceeding
MAXARTLINELENGTH and thus causing the article to be rejected.

Thanks to Florian Schlichting for the patch.


git-svn-id: http://inn.eyrie.org/svn/trunk@9086 38acf0ee-cff8-0310-b87c-e736704562cb