Add the possibility to run local innshellvars scripts.
innshellvars.local, innshellvars.pl.local and innshellvars.tcl.local
are now executed at the end of the run of the corresponding innshellvars script.
A typical use is to add or override variables.  For instance HOME or
GNUPGHOME.
These local scripts have to be executable; otherwise, they are not
taken into account.

Thanks to Matthew Vernon for the suggestion.


git-svn-id: http://inn.eyrie.org/svn/branches/2.5@9110 38acf0ee-cff8-0310-b87c-e736704562cb