Build INN with the new "-Wunused-but-set-variable" GCC 4.6.0 warning. It was complaining about an unused "sv" variable.
This commit fixes the following issues:
* innd and nnrpd segfaults because of a corrupted Perl stack.
A copy of the Perl stack pointer is saved at several places
in the code but not always restored the way it should be.  For instance,
PerlFilter() calls could reallocate the Perl stack and subsequently
cause failures when the stack is used afterwards.
Make sure that the Perl stack could not be modified twice in the row
without being properly updated at global scope.

* use ENTER-SAVETMPS-PUSHMARK-PUTBACK and SPAGAIN-PUTBACK-FREETMPS-LEAVE
macros whenever they are required.
Note that call_argv() does not need PUSHMARK-PUTBACK.

* fixed two memory leaks because of a xasprintf() which was not free'd.

* documentation of the filter_end() Perl routine that is called, if
defined, prior to the deactivation of Perl filters.  Both used by innd
and nnrpd when Perl filtering is turned off (either with "ctlinnd perl n"
or when a Perl function dies at run time).


git-svn-id: http://inn.eyrie.org/svn/trunk@9205 38acf0ee-cff8-0310-b87c-e736704562cb