better check of the rights on the private key for TLS connections
Use stat() instead of lstat().  The permissions of the symlink are
irrelevant; we only care about the permissions of the underlying file.

Use getegid() instead of getgid().

The news group is fairly trusted already by INN.  Ensure that if the mode
is 440 or 640, the group owner is the news group (to prevent the failure
case of having news:users as the owner and group).

In the error message, a wrong path to the key was put (cert_file instead
of key_file).

Thanks to Florian Schlichting for having reported the issue, and to Russ
Allbery for his help on the fix.


git-svn-id: http://inn.eyrie.org/svn/trunk@9219 38acf0ee-cff8-0310-b87c-e736704562cb