inncheck:  improvements in checks
* Allow uppercase characters in IPv6 addresses.
RFC 5952 section 4 / 4.3 says that all characters in an IPv6 address
MUST be textually represented in lowercase and advises that humans
should do so, too.  At the same time, all implementations must be able to
accept both lowercase and uppercase.

* Suppress (useless and misleading) error messages after an error
is detected, until the parser is back in sync.
The drawback is that it will hide two consecutive errors until the
first one gets fixed.

* Allow not to specify the maximum size in storage.conf.
"size: 16384," was treated as an error though it was totally valid
and meant "no upper limit".

* Recognize inclusions in readers.conf and innfeed.conf.
inncheck will ignore the include statement.
The drawback is that every file must be self-contained on its own:
a file must be a "complete" set of blocks and options.
In order to check the included file, inncheck has to be run again
on it (with for instance the option "readers.conf=/path/to/included/file").

Thanks to Florian Schlichting for the patch.


git-svn-id: http://inn.eyrie.org/svn/trunk@9230 38acf0ee-cff8-0310-b87c-e736704562cb