controlchan:  impose a date cutoff, block the replay of old control articles
controlchan now requires DateTime::Format::Mail to parse dates.
Otherwise, with the new Injection-Date: header field, old control articles
could be maliciously reinjected into Usenet, and replayed.

The Injection-Date: header of old control articles is not always signed...

controlchan now imposes a date cutoff.


According to RFC 5537, a relaying agent processes an article as follows:

  2.  It MUST examine the Injection-Date header field or, if absent,
      the Date header field, and reject the article if that date is
      more than 24 hours into the future.  It MAY reject articles with
      dates in the future with a smaller margin than 24 hours.

That is to say that the Date: header field can be set to anything when
an Injection-Date header field exists.


git-svn-id: http://inn.eyrie.org/svn/trunk@9238 38acf0ee-cff8-0310-b87c-e736704562cb