Build INN with the new "-Wunused-but-set-variable" GCC 4.6.0 warning.
Remove unused variables.

At the same time, do a bit of cleaning!  (Useless whitespaces,
useful (for reading) whitespaces, one-line if conditions,
"conn_ret" type instead of "int".)

Also add checks to WriteToWire_lmtpstr() results, so as to really
return the result (and properly disconnect the session on error),
instead of returning RET_OK as though everything were OK.

Same thing for WriteToWire_imapstr() at a few places.

No need to be in any auth step when we are about to
close the IMAP session after a failure of WriteToWire_imapstr().


git-svn-id: http://inn.eyrie.org/svn/branches/2.5@9252 38acf0ee-cff8-0310-b87c-e736704562cb