inncheck:  use Getopt::Long
Improve the way inncheck parses the command line arguments.  The last
matching argument is taken into account (consequently, "--noperm --perm
--noperm" is valid, as well as "-q -v").

As -perm is not just the opposite of -noperm, $perms is handled differently.
So 1 is -perm, 0 is -noperm, and -1 is neither.

Use --fix, --perms, etc. in our documentation.  Keep -fix, -perms, etc. for
backward compatibility.

Thanks to Florian Schlichting for this patch.


git-svn-id: http://inn.eyrie.org/svn/branches/2.5@9272 38acf0ee-cff8-0310-b87c-e736704562cb