improve checks on passwd.nntp
Improve the error message when the parsed lined is in wrong format:
"Argument list too long" (E2BIG) instead of "Numerical argument out
of domain" (EDOM).

Also catch a new kind of error in passwd.nntp:  any line prior to a matching
<hostname>, which is not empty or starts with a hash, but contains a third
colon somewhere that is not followed by "authinfo".

Thanks to Florian Schlichting for the patch.


git-svn-id: http://inn.eyrie.org/svn/branches/2.5@9273 38acf0ee-cff8-0310-b87c-e736704562cb