tradindexed remap of index files when writes out of order
The overview IDX files aren't being re-read after an article is written to
them -- the failing tests (in xref.t, in the test suite) are the places
where a stored overview line is read back to check that it was stored
correctly, except for the first time each group is read (tests 2-5, 11-18)
and when the high water mark is advanced (7-10), which work correctly.

There isn't any mechanism in the source code for the files to be refreshed
except in the case that the high water mark has been advanced.  On systems
which don't have MMAP_MISSES_WRITES, perhaps the underlying write shows
up automatically (?)

Thanks to Wim Lewis for the bug report.



git-svn-id: http://inn.eyrie.org/svn/branches/2.5@9279 38acf0ee-cff8-0310-b87c-e736704562cb