sendinpaths:  add a pattern for the dump filenames
Only process files whose name begin with "inpaths." for legacy reasons
(behaviour of the previous sendinpaths shell script).  It will prevent
issues on <pathlog>/path being symlinked to another directory.

Also abort the run of sendinpaths if we cannot chdir() to the log directory.

Thanks to The Doctor for having reported the problem.


git-svn-id: http://inn.eyrie.org/svn/trunk@9289 38acf0ee-cff8-0310-b87c-e736704562cb