add an INN::Utils::Shlock module to wrap shlock in Perl scripts
Calling shlock (a program shipped with INN) is more portable than using 
flock(2) and its corresponding Perl function because this function does
not work as expected on all existing systems (for instance on Solaris).

This patch:

* contains POD documentation for INN::Utils::Shlock;

* adds support for INN::Utils::Shlock instead of flock(2) in pullnews 
and mailpost.

* uses INN::Utils::Shlock instead of a subfunction calling shlock(1)
in controlchan, thdexpire and send-uucp.

Thanks to Dennis Davis for the bug report.


git-svn-id: http://inn.eyrie.org/svn/trunk@9303 38acf0ee-cff8-0310-b87c-e736704562cb