cnfsstat:  reload configuration files when modified
cnfsstat now reloads storage.conf and cycbuff.conf if they have been modified
since the last iteration of the script (started with the "-l" flag).
Otherwise, the new configuration is not taken into account by a "ctlinnd xexec"
command; only rc.news starts cnfsstat.

Also change the "Cannot open CycBuff Conffile" message because it is not
always true.  The message can be logged when an error occurs during the
parsing of the file.  Sometimes, even when the parsing was incorrect,
the file was accepted.  This behaviour is also fixed, and the file is
refused at cnfsstat startup.


git-svn-id: http://inn.eyrie.org/svn/trunk@9341 38acf0ee-cff8-0310-b87c-e736704562cb