docheckgroups:  no longer use awk
On a few systems, docheckgroups fails because of the use of an old
version of awk.  A few awk implementations have a limit in the size
of the input they can process; it then cause issues with large
newsgroups files.

Use Perl instead of awk for the time being.  It solves these issues.
Successfully tested on a system where docheckgroups was previously
failing.  Thanks to John F. Morse for the bug report.

The next step will be a total rewrite of docheckgroups.  It will
be much more maintainable as a Perl script.


git-svn-id: http://inn.eyrie.org/svn/trunk@9343 38acf0ee-cff8-0310-b87c-e736704562cb