add documentation for procbatch
New documentation written by Florian Schlichting, with the help of a memo
by Russ Allbery.

Add a link to this new documentation in innfeed and news.daily documentation.

Use "$INN::Config::pathtmp" instead of "INN::Config::$spooltmp" in procbatch.
(The two variables are the same; the first one is more wide-spread.)

Add a warning if -s is used without -c, and if -t is used without -m.


git-svn-id: http://inn.eyrie.org/svn/trunk@9371 38acf0ee-cff8-0310-b87c-e736704562cb