fixed a trailing extra junk byte in innd answers
When an error occurs when innd writes on a channel, a trailing extra junk
byte is added to the reply.

If the write() function fails, "i" is negative and we substract "i" to
cp->Out->left anyway.
So if for instance i=-1, we end up in having cp->Out->left increased
by one.  Which would add an extra character after a complete response.
(So it appears in fact before a new response.)

Issue introduced in changeset [7418].

Thanks to River Tarnell for the bug report.


git-svn-id: http://inn.eyrie.org/svn/trunk@9398 38acf0ee-cff8-0310-b87c-e736704562cb