Add additional dependency to de-parallelize innfeed yacc rule
With a highly parallel make, yacc could be invoked on the innfeed
configuration parser twice, once to build the *.c file and once to
build the *.h file, leading to one of the mv commands failing.
Follow the recommendation in the Automake manual and add an
additional dependency from the *.h file to the *.c file, forcing
the rule to be run once.

Reported by Remco Rijnders.


git-svn-id: http://inn.eyrie.org/svn/trunk@9451 38acf0ee-cff8-0310-b87c-e736704562cb