test suite:  remove a useless wrapped test for setenv
Remove a test that makes the test suite fail on a few systems.
The code it is testing is obvious whereas getting the amount of
memory and the data size limits right is extremely difficult
in all cases.  Besides, this test does not serve much purpose
and is not present in rra-c-util 4.7.


git-svn-id: http://inn.eyrie.org/svn/branches/2.5@9458 38acf0ee-cff8-0310-b87c-e736704562cb