fixed the occurrence of an unexpected "cant select" error generated by innd
errno was not preserved; the status hook was stuck between the select and
the check of its return status, so we're getting a spurious EPERM from
STATUSsummary, probably from some random glibc internal syscall involved
in fopen() that isn't actually important.

It only happens twice since it can only happen when select is interrupted
by a signal (triggering EINTR, which then gets overwritten with EPERM),
which is probably due to child processes completing, which is probably
only happening during specific events on your server during expire.

The solution is to rewrite this code to the order that it should have been
written in the first place.

Thanks to Paul Tomblin for having caught that long-standing issue.


git-svn-id: http://inn.eyrie.org/svn/trunk@9469 38acf0ee-cff8-0310-b87c-e736704562cb