innfeed, scanlogs:  reopen log files on sighup
The recent patch [9464] to flush exploder and process channels in innd,
and to send 'ctlinnd flushlogs' again after log rotation solved the
problem of controlchan or ninpaths holding open an already deleted
errlog file.  However this does not yet solve the original issue with
innfeed.log when innfeed is run as a standalone daemon via a funnel-file.

This patch implements the changes suggested by Julien Elie in January
2013, to add an openLogFile() call to the sighup handler.  The sigusr1
handler is not available, as it is already used to increase / decrease
the debugging level along with sigusr2.  Also, reopening the log after
re-reading the configuration might be a good idea anyway, as the log
file location may have been changed.

Additionally, call sighup on the innfeed pid in scanlogs, and document
the changes briefly in innfeed(8) / innfeed.conf(5).

Thanks to Florian Schlichting for the patch.


git-svn-id: http://inn.eyrie.org/svn/trunk@9484 38acf0ee-cff8-0310-b87c-e736704562cb