use at least Autoconf 2.64 to bootstrap INN
AC_INIT has been supporting 5 parameters since Autoconf 2.64.
Switch to this newer version.

Also improve HACKING as for the files taken from other projects when
releasing a new version of INN:
- mention the paths of these files in the libtool tarball,
- mention rra-c-util,
and document how to run a single test, as well as a pointer where
to find automatic build logs of INN.


git-svn-id: http://inn.eyrie.org/svn/trunk@9550 38acf0ee-cff8-0310-b87c-e736704562cb