update m4 files for Berkeley DB and zlib support
bdb.m4 and zlib.m4 are now part of rra-c-util.

The "--with-berkeleydb" configure flag used to add Berkeley DB support
is now "--with-bdb".

When building INN with Berkeley DB or zlib support, no longer add
standard locations to compiler and linker include flags.  Such default
paths are now added only if explicitly given to one or more of the
"--with-bdb", "--with-bdb-include", "--with-bdb-lib", "--with-zlib",
"--with-zlib-include", or "--with-zlib-lib" configure flags.  The flags
ending with "-include" and "-lib" are new.

If the Berkeley DB library is found at configure time, INN will now be
built with support for it unless the B"--without-bdb" flag is explicitly
passed to configure.
Note that it was already the default behaviour for zlib support when
Berkeley DB support was also enabled.

Use BDB_CPPFLAGS, BDB_LDFLAGS and BDB_LIBS instead of DB_CPPFLAGS,
DB_LDFLAGS and DB_LIBS.
Also use HAVE_BDB instead of USE_BERKELEY_DB.


git-svn-id: http://inn.eyrie.org/svn/trunk@9593 38acf0ee-cff8-0310-b87c-e736704562cb