nnrpd:  fixed an overlapping copy
Fixed an issue where users were denied posting because of an overlapping 
buffer copy in a check nnrpd was doing.

The error seen was "address not in Internet syntax" while using a From:
address of the form <a@test1.de>.  Debugging revealed that nnrpd copies
a buffer into itself to look at the part behind the "@", and for very
specific inputs on an old version of nnrpd that still uses strcpy, the
result did not contain the dot separating the top-level domain any more.

Thanks to Florian Schlichting for the patch.


git-svn-id: http://inn.eyrie.org/svn/trunk@9600 38acf0ee-cff8-0310-b87c-e736704562cb