improve documentation about using storeonxref with tradspool
Mention in the inn.conf man page that storeonxref needs being set to
true in case the tradspool article storage method is used.

Though a few parts of tradspool handle both values for storeonxref, the
implementation when storeonxref is false is not complete.  For instance,
CrackXref() is still called on the Newsgroups: header once, which
obviously fails.


git-svn-id: http://inn.eyrie.org/svn/trunk@9607 38acf0ee-cff8-0310-b87c-e736704562cb