Handle several UIDs on PGP keys when verifying control messages
Fixed a long-standing limitation on how controlchan and pgpverify
were checking the signer of control messages.  They now properly handle
the case of several UIDs being defined on a single PGP key, as well
as the presence of spaces into UIDs.  In previous versions of INN,
a few valid control messages got ignored because of that limitation
(fido.ger.* and grisbi.* were for instance impacted).


git-svn-id: http://inn.eyrie.org/svn/branches/2.5@9621 38acf0ee-cff8-0310-b87c-e736704562cb