Sync lib/buffer.c and include/inn/buffer.h with latest rra-c-util version
Changes are:

- Remove the bool arguments to buffer_sprintf and buffer_vsprintf and
instead introduce new buffer_append_sprintf and _vsprintf functions
to append to the buffer, which is what the functions did with a true
argument.  This avoids having a bool argument to functions, the meaning
of which is often obscure in the calling code.

- Fix buffer_free to support taking NULL pointers and doing nothing
with them.  Don't check whether a pointer is NULL before passing it
into free and instead assume free can handle NULL pointers properly.
This has been true for many years.

- Allocate memory with calloc and assume this sets pointers to NULL
instead of explicitly initializing them.  We already had to assume this
in various places, and architectures where the all-zero bit pattern is
not the NULL pointer are exceedingly rare.


git-svn-id: http://inn.eyrie.org/svn/trunk@9658 38acf0ee-cff8-0310-b87c-e736704562cb