Sync the messages test suite with upstream rra-c-util
- Use calloc in preference to calculating a malloc size with
multiplication everywhere.  In most places this caution was probably
not necessary, but uniformity is easier to audit and no one will ever
notice the speed difference between malloc and calloc.

- Add new message_handlers_reset function to the messages utility API.
This function resets all handlers to their defaults and frees any
memory allocated by the message_handlers functions.  This is primarily
useful to allow freeing all memory when doing exhaustive memory
allocation testing.

- Check the return status of vsnprintf in the syslog message handlers
for die, warn, and friends, and report an error with warn if vsnprintf
fails.

- Drop concat from the messages library.  asprintf or xasprintf, provided
by the portability and util libraries respectively, are entirely superior
alternatives to concat.

- Improve the is_function_output TAP add-on interface to take an opaque
data pointer and pass it into the called function.

- Add a new run_setup TAP add-on function that runs a given command and
calls bail if it fails, used for doing test setup in an external
command.  (Some things are easier to do in shell than in C.)


git-svn-id: http://inn.eyrie.org/svn/trunk@9683 38acf0ee-cff8-0310-b87c-e736704562cb