mailpost:  change the default location of the database and the temporary files
Add new -t flag to specify the default directory to use for temporary files.

Also change the default location of the persistent database from pathtmp
to patdb.  As a matter of fact, mailpost tries to store its Message-ID
database in INN's pathtmp, but since it's usually running as the mail
system (often daemon) rather than as news, this fails.

Check that the directories are writable when mailpost is run, and
otherwise die with an error.

close #81


git-svn-id: http://inn.eyrie.org/svn/trunk@9830 38acf0ee-cff8-0310-b87c-e736704562cb